requirements.txt
Flask
openai
requests
python-dotenv
